[Feat] 정보보기 페이지 UI 제작
