+++ src/App.js
@@ -3,9 +3,10 @@ import { Route, Routes } from "react-router-dom";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
 import Main from "./pages/Main";
-import View from "./pages/View";
 import Share from "./pages/Share";
 import Together from "./pages/Together";
+import PerformanceDetail from "./pages/PerformanceDetail"
+import FestivalDetail from "./pages/FestivalDetail"
 
 function App() {
   return (
@@ -14,9 +15,11 @@ function App() {
       <Header />
       <Routes>
         <Route path="/" element={<Main />} />
-        <Route path="/view" element={<View />} />
         <Route path="/share" element={<Share />} />
         <Route path="/together" element={<Together />} />
+        <Route path="/view/performance_detail" element={<PerformanceDetail />} />
+        <Route path="/view/festival_detail" element={<FestivalDetail />} />
+
       </Routes>
       <Footer />
     </>
